refactor(db): extract splitSQLStatements so it can be tested
db-manager.js is a CLI that connects to a database as a side effect of
being imported, so the splitter could not be reached from a test while it
lived there. Moved verbatim into its own module, no behaviour change.
